Do not expose LibGit2Sharp usage outside of the NuKeeper.Git project
Minor cleanup, might be useful if some replacement git driver is considered in order to support shallow checkout.
