+++ NuKeeper.Abstractions/Git/GitUsernamePasswordCredentials.cs
@@ -1,0 +1,9 @@
+namespace NuKeeper.Abstractions.Git
+{
+    public class GitUsernamePasswordCredentials
+    {
+        public string Username { get; set; }
+
+        public string Password { get; set; }
+    }
+}
